Moved noise.add_noise_to_program back to kraus.add_noise_to_program
