+++ pyquil/noise.py
@@ -16,14 +16,12 @@
 """
 Module for creating and verifying noisy gate and readout definitions.
 """
-import warnings
 from collections import namedtuple
 
 import numpy as np
 
 from pyquil.parameters import format_parameter
-from pyquil.quilbase import Pragma, Gate, Qubit as QuilQubit
-
+from pyquil.quilbase import Pragma, Gate
 
 INFTY = float("inf")
 "Used for infinite coherence times."
@@ -470,35 +468,6 @@ def apply_noise_model(prog, noise_model):
         else:
             new_prog += i
     return new_prog
-
-
-def add_noise_to_program(prog, T1=30e-6, T2=None, gate_time_1q=50e-9, gate_time_2q=150e-09,
-                         ro_fidelity=0.95):
-    """
-    Add generic damping and dephasing noise to a program.
-
-    .. warning::
-
-        This function is deprecated. Please use :py:func:`add_decoherence_noise` instead.
-
-    :param prog: A pyquil program consisting of I, RZ, CZ, and RX(+-pi/2) instructions
-    :param Union[Dict[int,float],float] T1: The T1 amplitude damping time either globally or in a
-        dictionary indexed by qubit id. By default, this is 30 us.
-    :param Optional[Union[Dict[int,float],float]] T2: The T2 dephasing time either globally or in a
-        dictionary indexed by qubit id. By default, this is one-half of the T1 time.
-    :param float gate_time_1q: The duration of the one-qubit gates, namely RX(+pi/2) and RX(-pi/2).
-        By default, this is 50 ns.
-    :param float gate_time_2q: The duration of the two-qubit gates, namely CZ.
-        By default, this is 150 ns.
-    :param Union[Dict[int,float],float] ro_fidelity: The readout assignment fidelity
-        :math:`F = (p(0|0) + p(1|1))/2` either globally or in a dictionary indexed by qubit id.
-    :return: A new program with noisy operators.
-    """
-    warnings.warn("pyquil.noise.add_noise_to_program is deprecated, please use "
-                  "pyquil.noise.add_decoherence_noise instead.",
-                  DeprecationWarning)
-    return add_decoherence_noise(prog, T1=T1, T2=T2, gate_time_1q=gate_time_1q,
-                                 gate_time_2q=gate_time_2q, ro_fidelity=ro_fidelity)
 
 
 def add_decoherence_noise(prog, T1=30e-6, T2=None, gate_time_1q=50e-9, gate_time_2q=150e-09,
